Organize the Computest plugin into separate modules
